projects/phone_app/requirements.txt
customtkinter==5.2.2
pylance==0.13.0
pillow==10.3.0